Refactor: apis, hooks 폴더 리팩토링 #41
