+++ src/hooks/auth/useIsLoggedIn.tsx
@@ -1,9 +1,8 @@
 import {useEffect, useState} from 'react';
 import {useSelector} from 'react-redux';
-import {RootState} from '../redux/stores/reducer';
+import {RootState} from '../../redux/stores/reducer.ts';
 
-// redux의 userSlice에서 accessToken && refreshToken 일 때를
-// login 상태로 간주
+// redux의 userSlice에서 accessToken && refreshToken 일 때를 login 상태로 간주
 
 const useIsLoggedIn = () => {
   const [isLoggedIn, setIsLoggedIn] = useState<boolean>(false);
@@ -15,8 +14,10 @@ const useIsLoggedIn = () => {
 
   useEffect(() => {
     if (accessToken && refreshToken) {
+      console.log('📝User is logged in.');
       setIsLoggedIn(true);
     } else {
+      console.log('✅User is logged Off.');
       setIsLoggedIn(false);
     }
   }, [accessToken, refreshToken]);
