+++ src/hooks/auth/useLogout.tsx
@@ -1,8 +1,9 @@
 import {useState, useCallback} from 'react';
-import {useAppDispatch} from '../redux/stores';
-import userSlice from '../redux/slices/user';
-import logout from '../apis/auth/logout';
+import {useAppDispatch} from '../../redux/stores';
+import userSlice from '../../redux/slices/user.ts';
+import logout from '../../apis/auth/logout.ts';
 import EncryptedStorage from 'react-native-encrypted-storage';
+import useIsLoggedIn from './useIsLoggedIn.tsx';
 
 const useLogout = () => {
   const dispatch = useAppDispatch();
