+++ src/hooks/login/useKakaoLogin.tsx
@@ -1,10 +1,10 @@
 import {useState} from 'react';
 import {getAccessToken} from '@react-native-seoul/kakao-login';
-import loginSocial from '../apis/auth/loginSocial.ts';
-import {useAppDispatch} from '../redux/stores';
+import loginSocial from '../../apis/auth/loginSocial.ts';
+import {useAppDispatch} from '../../redux/stores';
 import EncryptedStorage from 'react-native-encrypted-storage';
-import userSlice from '../redux/slices/user.ts';
-import useSetAccessTokenAndRefreshToken from './useSetAccessTokenAndRefreshToken.ts';
+import userSlice from '../../redux/slices/user.ts';
+import useSetAccessTokenAndRefreshToken from '../auth/useSetAccessTokenAndRefreshToken.ts';
 import {useNavigation} from '@react-navigation/native';
 
 // useKakaoLogin 커스텀 훅 수정
