+++ src/hooks/popUpList/useGetNewList.tsx
@@ -1,15 +1,15 @@
 import {useState, useEffect} from 'react';
-import getHotList from '../apis/popup/hotList.ts';
-import {GetPopUpListResponse} from '../types/PopUpListData.ts';
+import {GetPopUpListResponse} from '../../types/PopUpListData.ts';
+import getNewList from '../../apis/popup/newList.ts';
 
-interface HotListState {
+interface GetNewState {
   loading: boolean;
   error: Error | null;
   data: GetPopUpListResponse[] | null;
 }
 
-const useGetHotList = () => {
-  const [hotListState, setHotListState] = useState<HotListState>({
+const useGetNewList = () => {
+  const [getListState, setGetListState] = useState<GetNewState>({
     loading: false,
     error: null,
     data: null,
@@ -17,20 +17,20 @@ const useGetHotList = () => {
 
   useEffect(() => {
     const fetchHotList = async () => {
-      setHotListState(prevState => ({...prevState, loading: true}));
+      setGetListState(prevState => ({...prevState, loading: true}));
       try {
-        const response = await getHotList();
+        const response = await getNewList();
         if (response.success) {
-          setHotListState({loading: false, error: null, data: response.data!}); // API 응답에 따라 'data'나 'result'로 접근 가능
+          setGetListState({loading: false, error: null, data: response.data});
         } else {
-          setHotListState({
+          setGetListState({
             loading: false,
             error: new Error(response.error?.message || 'Unknown error'),
             data: null,
           });
         }
       } catch (error: any) {
-        setHotListState({
+        setGetListState({
           loading: false,
           error:
             error instanceof Error
@@ -44,7 +44,7 @@ const useGetHotList = () => {
     fetchHotList();
   }, []);
 
-  return hotListState;
+  return getListState;
 };
 
-export default useGetHotList;
+export default useGetNewList;
